# docs/openapi-identity.yaml
openapi: 3.1.0
info:
  title: Pulso Piura API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /reservations:
    post:
      summary: Crea un hold idempotente para una franja vigente
      description: El servidor deriva organización, precio, moneda y vencimiento y limita las reservas temporales activas por cliente.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, minLength: 1, maxLength: 100}
        - in: header
          name: X-Correlation-Id
          required: false
          schema: {type: string, maxLength: 100}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateReservationInput"}
      responses:
        "201": {description: Hold creado, content: {application/json: {schema: {$ref: "#/components/schemas/Reservation"}}}}
        "200": {description: Reintento idempotente; devuelve la reserva creada anteriormente, content: {application/json: {schema: {$ref: "#/components/schemas/Reservation"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {description: Token ausente o inválido}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {description: La franja no coincide con una regla vigente o pertenece al pasado}
        "429": {description: El cliente alcanzó el máximo de reservas temporales activas}
  /me/reservations:
    get:
      summary: Lista las reservas del usuario autenticado
      security: [{ bearerAuth: [] }]
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 0, default: 0}}
        - {in: query, name: size, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        "200": {description: Historial propio paginado, content: {application/json: {schema: {$ref: "#/components/schemas/ReservationPage"}}}}
        "401": {description: Token ausente o inválido}
  /reservations/{reservationId}:
    parameters:
      - {in: path, name: reservationId, required: true, schema: {type: string, format: uuid}}
    get:
      summary: Consulta una reserva propia o de una organización autorizada
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Reserva visible, content: {application/json: {schema: {$ref: "#/components/schemas/Reservation"}}}}
        "401": {description: Token ausente o inválido}
        "404": {$ref: "#/components/responses/NotFound"}
  /reservations/{reservationId}/confirm:
    parameters:
      - {in: path, name: reservationId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Confirma idempotentemente un hold vigente sin adelanto
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Reserva confirmada, content: {application/json: {schema: {$ref: "#/components/schemas/Reservation"}}}}
        "401": {description: Token ausente o inválido}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /reservations/{reservationId}/cancel:
    parameters:
      - {in: path, name: reservationId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Cancela idempotentemente una reserva gestionable
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Reserva cancelada, content: {application/json: {schema: {$ref: "#/components/schemas/Reservation"}}}}
        "401": {description: Token ausente o inválido}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /reservations/{reservationId}/check-in-pass:
    parameters:
      - {in: path, name: reservationId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Emite o rota el pase QR opaco de una reserva propia confirmada
      description: El valor anterior queda invalidado y el servidor almacena únicamente su hash SHA-256.
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Pase emitido, content: {application/json: {schema: {$ref: "#/components/schemas/ReservationCheckInPass"}}}}
        "401": {description: Token ausente o inválido}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/reservations/check-in/preview:
    parameters:
      - {in: path, name: organizationId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Valida un QR y muestra la reserva y su saldo al dueño del complejo
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReservationCheckInInput"}
      responses:
        "200": {description: Reserva válida, content: {application/json: {schema: {$ref: "#/components/schemas/ReservationCheckIn"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/reservations/check-in:
    parameters:
      - {in: path, name: organizationId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Registra idempotentemente la llegada y completa la reserva
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReservationCheckInInput"}
      responses:
        "200": {description: Llegada registrada, content: {application/json: {schema: {$ref: "#/components/schemas/ReservationCheckIn"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /platform/organizations:
    get:
      summary: Lista organizaciones con sus dueños activos
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Organizaciones y dueños}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/organizations/{organizationId}/owners:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    post:
      summary: Asigna como dueño a una cuenta registrada
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Dueño asignado}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/organizations/{organizationId}/owners/{userId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - {in: path, name: userId, required: true, schema: {type: string, format: uuid}}
    delete:
      summary: Retira un dueño conservando al menos uno activo
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Dueño retirado}
        "409": {description: No se puede retirar al único dueño}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/businesses:
    get:
      summary: Lista todos los negocios para administración
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Negocios registrados}
        "403": {description: Requiere PLATFORM_ADMIN}
    post:
      summary: Crea un negocio aliado
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/PartnerBusinessInput"}}}
      responses:
        "201": {description: Negocio creado}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/businesses/{businessId}:
    parameters:
      - {in: path, name: businessId, required: true, schema: {type: string, format: uuid}}
    put:
      summary: Actualiza o cambia la publicación de un negocio
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/PartnerBusinessInput"}}}
      responses:
        "200": {description: Negocio actualizado}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {description: Requiere PLATFORM_ADMIN}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId: {in: path, name: organizationId, required: true, schema: {type: string, format: uuid}}
...
  responses:
    BadRequest: {description: Entrada inválida, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Forbidden: {description: Sin membresía o permiso en la organización, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    NotFound: {description: Recurso no encontrado dentro del tenant, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Conflict: {description: Estado, versión o dato en conflicto, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
...
    CreateReservationInput:
      type: object
      required: [sportSpaceId, startsAt, endsAt]
      properties:
        sportSpaceId: {type: string, format: uuid}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
    Reservation:
      type: object
      required: [id, sportSpaceId, startsAt, endsAt, status, totalMinor, depositMinor, currency, version]
      properties:
        id: {type: string, format: uuid}
        sportSpaceId: {type: string, format: uuid}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        status: {type: string, enum: [HOLD, PENDING_PAYMENT, CONFIRMED, COMPLETED, CANCELLED, EXPIRED]}
        totalMinor: {type: integer, format: int64, minimum: 0}
        depositMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, const: PEN}
        expiresAt: {type: [string, "null"], format: date-time}
        version: {type: integer, format: int64, minimum: 0}
    ReservationPage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Reservation"}}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 50}
        total: {type: integer, format: int64, minimum: 0}
    ReservationCheckInPass:
      type: object
      required: [reservationId, payload, issuedAt, validUntil, reservationVersion]
      properties:
        reservationId: {type: string, format: uuid}
        payload: {type: string, pattern: '^PULSO-CHECKIN:[A-Za-z0-9_-]{43}$'}
        issuedAt: {type: string, format: date-time}
        validUntil: {type: string, format: date-time}
        reservationVersion: {type: integer, format: int64, minimum: 0}
    ReservationCheckInInput:
      type: object
      required: [payload]
      properties:
        payload: {type: string, minLength: 43, maxLength: 100}
    ReservationCheckIn:
      type: object
      required: [reservationId, venueName, spaceName, customerName, startsAt, endsAt, reservationStatus, totalMinor, paidMinor, balanceMinor, currency, paymentStatus]
      properties:
        reservationId: {type: string, format: uuid}
        venueName: {type: string}
        spaceName: {type: string}
        customerName: {type: string}
        customerEmail: {type: [string, "null"], format: email}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        reservationStatus: {type: string, enum: [CONFIRMED, COMPLETED]}
        totalMinor: {type: integer, format: int64, minimum: 0}
        paidMinor: {type: integer, format: int64, minimum: 0}
        balanceMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, const: PEN}
        paymentStatus: {type: string, enum: [PAID, PARTIAL, PENDING]}
        checkedInAt: {type: [string, "null"], format: date-time}
...
    PartnerBusinessInput:
      type: object
      required: [name, category, zone, status]
      properties:
        name: {type: string, maxLength: 160}
        category: {type: string, enum: [CHOPERIA, RESTAURANT, SPORTS_BAR, OTHER]}
        zone: {type: string, maxLength: 180}
        description: {type: [string, "null"], maxLength: 500}
        imageUrl: {type: [string, "null"], maxLength: 500}
        contactPhone: {type: [string, "null"], maxLength: 30}
        latitude: {type: [number, "null"], minimum: -90, maximum: 90}
        longitude: {type: [number, "null"], minimum: -180, maximum: 180}
        status: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
...